Workaround menu click failure
--- Why

Margarita witnessed failure to scroll to a section upon clicking a menuitem
immediately after pageload.  I've been unable to reproduce.  The click handler
is very simple, so it's not dying inside the handler.  The handler function
should be defined before the menu becomes visible.

--- What

As an attempted workaround, I'll reintroduce link tags for the menu items so
that if the handler doesn't get called, the browser link default behavior will
jump to the target section.